requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
joblib==0.13.2
numpy==1.17.2
Pillow==6.1.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
urllib3==1.25.3
